Improved debugging and fixed problem with RemoveBaseline_nOrder
Implemented a 'debugger' class, that handles all the plotting for debugging. Can be passed to functions instead of the previous 'debug = True' to allow them to plot data etc.

Changed the rejection threshold in RemoveBaseline_nOrder to be relative to the number of datapoints in the approach region selected, so it no longer rejects curves with fast/short flat regions.